refactor: remove guard in a new module
